+++ AngularClient/projects/backbone-client-core/src/lib/settings.ts
@@ -1,9 +1,13 @@
 export interface ISettings {
+  environment: string;
   serviceApiUrl: string;
   serviceODataUrl: string;
+  sourceMapMappingsUrl: string;
 }
 
 export class Settings implements ISettings {
+  environment = "";
   serviceApiUrl = "";
   serviceODataUrl = "";
+  sourceMapMappingsUrl = "";
 }